models/summarization/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
torch==2.0.0
transformers==4.30.0
PyPDF2==3.0.1
python-docx==0.8.11
sentencepiece==0.1.99
numpy>=1.19.5 